+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2023-02-22
+
 ### Added
 
 - Support for labels and content type, [PR-50](https://github.com/reductstore/reduct-cpp/pull/50)
@@ -123,7 +125,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation of Reduct Storage API v0.1.0
 
-[Unreleased]: https://github.com/reduct-storage/reduct-cpp/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/reduct-storage/reduct-cpp/compare/v1.3.0...HEAD
+
+[1.3.0]: https://github.com/reduct-storage/reduct-cpp/compare/v1.2.0...1.3.0
 
 [1.2.0]: https://github.com/reduct-storage/reduct-cpp/compare/v1.1.0...1.2.0